refactor: migrate use-cache-transform to optional dependency with lazy loading
- Move @rari/use-cache-transform from dependencies to optionalDependencies in rari package.json
- Remove quick-lru dependency and inline cache implementation
- Create use-cache-loader.ts to lazily load use-cache-transform module at runtime
- Update Vite plugin transform hooks to use async getUseCacheTransform() instead of direct imports
- Re-export cache utilities from @rari/use-cache-transform in rari's cache-wrapper.ts
- Move cache implementation (cache-wrapper, deterministic-stringify, transform) into use-cache-transform package
- Add TypeScript and Vite configuration to use-cache-transform package
- Update test references to use new package structure
- Enables optional native module support while maintaining fallback behavior
