+++ packages/rari/package.json
@@ -122,13 +122,11 @@
     "react-dom": "catalog:react"
   },
   "dependencies": {
-    "@rari/use-cache-transform": "workspace:*",
     "lightningcss": "catalog:build",
-    "quick-lru": "catalog:",
-
     "rolldown": "catalog:build"
   },
   "optionalDependencies": {
+    "@rari/use-cache-transform": "workspace:*",
     "rari-darwin-arm64": "0.14.12",
     "rari-darwin-x64": "0.14.12",
     "rari-linux-arm64": "0.14.12",
@@ -140,7 +138,6 @@
     "@rari/deploy": "workspace:*",
     "@rari/logger": "workspace:*",
     "@types/node": "catalog:typescript",
-
     "@types/react": "catalog:react",
     "@types/react-dom": "catalog:react",
     "@typescript/native-preview": "catalog:typescript",
