--- packages/use-cache-transform/index.d.ts
@@ -1,30 +1,0 @@
-export interface TransformOptions {
-  filename: string
-  hashSalt?: string
-  cacheKinds?: string[]
-}
-
-export interface TransformResult {
-  code: string
-  needsReactCache: boolean
-  needsCacheWrapper: boolean
-  needsRegisterRef: boolean
-}
-
-export interface NativeAddon {
-  detectUseCache: (source: string) => boolean
-  transformUseCache: (
-    source: string,
-    options: TransformOptions,
-  ) => TransformResult
-}
-
-declare const nativeAddon: NativeAddon
-
-export function detectUseCache(source: string): boolean
-export function transformUseCache(
-  source: string,
-  options: TransformOptions,
-): TransformResult
-
-export default nativeAddon
