+++ packages/use-cache-transform/src/index.ts
@@ -1,0 +1,8 @@
+export { detectUseCache, transformUseCache } from './native'
+export type { NativeAddon, TransformOptions, TransformResult } from './native'
+
+export { $$cache__, encodeBoundArgs } from './runtime/cache-wrapper'
+export type { CacheEntry } from './runtime/cache-wrapper'
+
+export { transformUseCacheModule } from './transform'
+export type { UseCacheTransformOptions } from './transform'
