+++ packages/use-cache-transform/src/transform.ts
@@ -1,5 +1,5 @@
-import type { NativeAddon } from '@rari/use-cache-transform'
-import nativeAddon, { transformUseCache } from '@rari/use-cache-transform'
+import type { NativeAddon } from './native'
+import nativeAddon, { transformUseCache } from './native'
 
 const USE_CACHE_FUNCTION_REGEX = /['"]use\s+cache(?::\s*[\w-]+)?['"]/
 const DIRECTIVE_PROLOGUE_REGEX = /^['"][^'"]+['"];?\s*$/
@@ -12,11 +12,15 @@ let addon: NativeAddon | null = null
 let addonLoadAttempted = false
 
 function getAddon(): NativeAddon | null {
-  if (addonLoadAttempted) {
+  if (addonLoadAttempted)
     return addon
-  }
+
   addonLoadAttempted = true
   addon = nativeAddon
+
+  if (!addon)
+    console.warn('[use-cache-transform] Native addon not available — transforms will be skipped')
+
   return addon
 }
 
@@ -29,21 +33,27 @@ function extractPrologueLines(code: string): string[] {
       prologue.push(line)
       continue
     }
-    if (DIRECTIVE_PROLOGUE_REGEX.test(trimmed)) {
+    if (DIRECTIVE_PROLOGUE_REGEX.test(trimmed))
       prologue.push(line)
-    }
-    else {
+    else
       break
-    }
   }
 
   return prologue
 }
 
-export function transformUseCacheModule(code: string, id: string): string | null {
-  if (!hasUseCacheFunction(code)) {
+export interface UseCacheTransformOptions {
+  hashSalt?: string
+  cacheKinds?: string[]
+}
+
+export function transformUseCacheModule(
+  code: string,
+  id: string,
+  options: UseCacheTransformOptions = {},
+): string | null {
+  if (!hasUseCacheFunction(code))
     return null
-  }
 
   console.error(`[use-cache-transform] found 'use cache' directive in ${id}`)
 
@@ -56,8 +66,8 @@ export function transformUseCacheModule(code: string, id: string): string | null
   try {
     const result = transformUseCache(code, {
       filename: id,
-      hashSalt: 'rari-use-cache-v1',
-      cacheKinds: ['default'],
+      hashSalt: options.hashSalt || 'rari-use-cache-v1',
+      cacheKinds: options.cacheKinds || ['default'],
     })
 
     if (result.code === code) {
@@ -68,17 +78,14 @@ export function transformUseCacheModule(code: string, id: string): string | null
     console.error(`[use-cache-transform] transformed ${id} (needsCacheWrapper=${result.needsCacheWrapper}, needsRegisterRef=${result.needsRegisterRef})`)
 
     const imports = []
-    if (result.needsReactCache) {
+    if (result.needsReactCache)
       imports.push(`import { cache as $$reactCache__ } from 'react'`)
-    }
 
-    if (result.needsCacheWrapper) {
+    if (result.needsCacheWrapper)
       imports.push(`import { $$cache__, encodeBoundArgs } from 'rari/runtime/cache-wrapper'`)
-    }
 
-    if (result.needsRegisterRef) {
+    if (result.needsRegisterRef)
       imports.push(`import { registerServerReference } from 'rari/runtime/react-server-dom-shim'`)
-    }
 
     const prologueLines = extractPrologueLines(result.code)
     const importBlock = imports.length ? `${imports.join(';\n')};\n` : ''
