+++ packages/use-cache-transform/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../.config/typescript/package.json",
+  "include": ["src"]
+}
